+++ WhateverGreen/kern_weg.hpp
@@ -12,7 +12,7 @@
 #include <Headers/kern_devinfo.hpp>
 
 #include "kern_rad.hpp"
-
+#include "kern_shiki.hpp"
 
 class IOFramebuffer;
 class IODisplay;
@@ -40,8 +40,8 @@ private:
 	/**
 	 *  Hardware acceleration and FairPlay fixes instance
 	 */
+	SHIKI shiki;
 	
-
 	/**
 	 *  FB_DETECT   autodetects based on the installed GPU.
 	 *  FB_RESET    enforces -v like usual patch.
@@ -75,9 +75,19 @@ private:
 	};
 
 	/**
-	 *  applbkkl boot-arg controlled AppleBacklight kext patch
+	 *  applbkl boot-arg controlled AppleBacklight kext patch
 	 */
 	uint32_t appleBacklightPatch {APPLBKL_DETECT};
+
+	/**
+	 *  applbkl custom device name if any
+	 */
+	OSData *appleBacklightCustomName {nullptr};
+
+	/**
+	 *  applbkl custom device data if any
+	 */
+	OSData *appleBacklightCustomData {nullptr};
 
 	/**
 	 *  Backlight panel data format